requirements.txt
click
boto3
requests
GitPython